+++ Maze/APPAppDelegate.m
@@ -27,6 +27,7 @@
 
 - (void)applicationDidEnterBackground:(UIApplication *)application
 {
+    [self.viewController stopMotionUpdates];
     // Use this method to release shared resources, save user data, invalidate timers, and store enough application state information to restore your application to its current state in case it is terminated later. 
     // If your application supports background execution, this method is called instead of applicationWillTerminate: when the user quits.
 }
@@ -44,6 +45,7 @@
 
 - (void)applicationWillTerminate:(UIApplication *)application
 {
+    [self.viewController stopMotionUpdates];
     // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
 }