+++ Maze/APPMotionValidation.h
@@ -3,6 +3,8 @@
 
 #include <stdbool.h>
 
-bool APPMotionComponentsAreFinite(double x, double y, double z);
+#define APPMotionMaximumAccelerationComponent 16.0
+
+bool APPMotionComponentsAreSafeForIntegration(double x, double y, double z);
 
 #endif